esta separado agora os clientLOG e parte de verificar solucao atraves da outra solucao ns bem se era isso apenas

+++ cliente.h
@@ -7,7 +7,7 @@ void imprimirTabuleiro(char* jogo);
 void carregarConfigCliente(char* nomeFicheiro);
 
 struct ClienteConfig{
-    int idCliente;
+    unsigned long idCliente;
     char ipServidor[IP_SIZE];
 };
